Single header file for config.
