--- orm_lib/inc/drogon/orm/config/Sqlite3Config.h
@@ -1,13 +1,0 @@
-#pragma once
-#include <string>
-
-namespace drogon::orm
-{
-struct Sqlite3Config
-{
-    size_t connectionNumber;
-    std::string filename;
-    std::string name;
-    double timeout;
-};
-}  // namespace drogon::orm
